Refactor password fields
